# docs/api/openapi.yaml
openapi: "3.1.0"
info:
  title: Supply Chain AI Platform API
  description: |
    Production-grade multi-agent AI supply chain orchestration API powered by Claude Opus 4.6.

    ## Features
    - **Multi-Agent Orchestration**: 7 specialized AI agents coordinated by Claude Opus 4.6
    - **Real-time Inventory Management**: Stock tracking with automated reorder recommendations
    - **Order Processing**: End-to-end order lifecycle with fraud detection
    - **Supplier Management**: Vendor evaluation and RFQ generation
    - **Demand Forecasting**: ML-based predictions using XGBoost
    - **Anomaly Detection**: Real-time fraud and anomaly scoring

    ## Authentication
    All endpoints require JWT Bearer token authentication (except `/health`, `/ready`, and `/api/v1/auth/token`).

    ```
    Authorization: Bearer <your-jwt-token>
    ```

    Obtain tokens via `POST /api/v1/auth/token` (OAuth2 password flow).

    ## Rate Limiting
    - 100 requests/minute per authenticated user
    - 1000 requests/hour per authenticated user
    - Agent orchestration: 10 requests/minute per user
  version: "1.0.0"
  contact:
    name: Platform Team
    email: platform@supply-chain.internal
  license:
    name: Proprietary
    url: https://supply-chain.internal/license
servers:
  - url: http://localhost:8000
    description: Local Development
  - url: https://api.supply-chain.internal
    description: Production
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/inventory:
    get:
      tags: [Inventory]
      summary: List inventory items
      operationId: list_inventory
      parameters:
        - name: page
          in: query
          schema: {type: integer, default: 1}
        - name: page_size
          in: query
          schema: {type: integer, default: 20}
        - name: category
          in: query
          schema: {type: string}
        - name: low_stock_only
          in: query
          schema: {type: boolean, default: false}
          description: Return only items at or below reorder point
      responses:
        "200":
          description: Inventory list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryListResponse"

    post:
      tags: [Inventory]
      summary: Create inventory item
      operationId: create_inventory_item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryCreate"
      responses:
        "201":
          description: Inventory item created
...
  /api/v1/inventory/{sku}:
    get:
      tags: [Inventory]
      summary: Get inventory item by SKU
      operationId: get_inventory_item
      parameters:
        - name: sku
          in: path
          required: true
          schema: {type: string}
      responses:
        "200":
          description: Inventory item details
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/inventory/{sku}/adjust:
    post:
      tags: [Inventory]
      summary: Adjust inventory quantity
      operationId: adjust_inventory
      parameters:
        - name: sku
          in: path
          required: true
          schema: {type: string}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryAdjustment"
      responses:
        "200":
          description: Inventory adjusted
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
...
    InventoryCreate:
      type: object
      required: [sku, name, unit_cost]
      properties:
        sku:
          type: string
        name:
          type: string
        category:
          type: string
        unit_cost:
          type: number
          minimum: 0
        unit_price:
          type: number
        quantity_on_hand:
          type: integer
          minimum: 0
          default: 0
        reorder_point:
          type: integer
          default: 10
        reorder_quantity:
          type: integer
          default: 50
        warehouse_location:
          type: string
    InventoryAdjustment:
      type: object
      required: [quantity_delta, reason]
      properties:
        quantity_delta:
          type: integer
          description: "Positive to add, negative to remove"
        reason:
          type: string
          enum: [purchase, sale, damaged, returned, audit, transfer]
        notes:
          type: string
    InventoryListResponse:
      type: object
      properties:
        items:
          type: array
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
...